fix tests for file exists
